+++ agent/go-service/pretask/gamesetting/regedit_other.go
