+++ agent/go-service/pretask/gamesetting/regedit_windows.go
