+++ agent/go-service/taskersink/aspectratio/checker.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/MaaXYZ/MaaEnd/agent/go-service/exetask/gamesetting"
+	"github.com/MaaXYZ/MaaEnd/agent/go-service/pretask/gamesetting"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/pkg/control"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/pkg/i18n"
 	"github.com/MaaXYZ/MaaEnd/agent/go-service/pkg/maafocus"
